Added monitoring for the time of the last action that occurred.
There are two new setter & getter functions in the Indicators section that tell me when was the time when the last pneumatic action took place. This is useful for creating a better autooff feature. The setter allows manually to tell the system to take the present time as the time of last activity.